% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/20.0_ToxCastSummary_finalFunction.R
\name{Assay_comparison}
\alias{Assay_comparison}
\title{Function for AC50 comparison of VAMR endpoints with available toxcast assays}
\usage{
Assay_comparison(vamr_file, cytotox_lb, cytotox_median, pathToExp, do_filter)
}
\arguments{
\item{vamr_file:}{outputfile from concnetration-response modeling containing all endpoints of the chemical of interest}

\item{cytotox_lb:}{This is the value of cytotox}
}
\description{
Function for AC50 comparison of VAMR endpoints with available toxcast assays
}
