% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_logsheet_adaption_func.R
\name{logsheet_adapt}
\alias{logsheet_adapt}
\title{Standardize logsheets}
\usage{
logsheet_adapt(pathToExp)
}
\arguments{
\item{pathToExp}{Character string. Path to the experiment directory containing necessary subdirectories for input and output}
}
\description{
This function can transform old logsheets to the new standardized logsheet format.
}
